move all emmet.cli to emmet.cli.legacy
