+++ .pre-commit-config.yaml
@@ -7,7 +7,7 @@ repos:
     rev: v0.0.261
     hooks:
       - id: ruff
-        args: [--fix, --ignore, "D,E501,E741,E402", --exclude, "__init__.py,emmet-cli,*_resources.py"]
+        args: [--fix, --ignore, "D,E501,E741,E402", --exclude, "__init__.py,emmet-cli-legacy,*_resources.py"]
 
   - repo: https://github.com/psf/black
     rev: 23.3.0
