+++ emmet-cli-legacy/emmet/cli/legacy/admin.py
@@ -3,8 +3,8 @@ import logging
 import click
 from pymatgen.core import Structure
 
-from emmet.cli import SETTINGS
-from emmet.cli.utils import EmmetCliError, ensure_indexes, get_meta_from_structure
+from emmet.cli.legacy import SETTINGS
+from emmet.cli.legacy.utils import EmmetCliError, ensure_indexes, get_meta_from_structure
 
 logger = logging.getLogger("emmet")
