+++ emmet-cli-legacy/emmet/cli/legacy/decorators.py
@@ -10,8 +10,8 @@ import click
 from slurmpy import Slurm
 from github3.gists import ShortGist
 
-from emmet.cli import SETTINGS
-from emmet.cli.utils import EmmetCliError, ReturnCodes, reconstruct_command
+from emmet.cli.legacy import SETTINGS
+from emmet.cli.legacy.utils import EmmetCliError, ReturnCodes, reconstruct_command
 
 logger = logging.getLogger("emmet")
 COMMENT_TEMPLATE = """
