+++ emmet-cli-legacy/setup.py
@@ -3,7 +3,7 @@ import datetime
 from setuptools import setup
 
 setup(
-    name="emmet-cli",
+    name="emmet-cli-legacy",
     version=datetime.datetime.today().strftime("%Y.%m.%d"),
     description="command line interface for emmet",
     author="The Materials Project",
@@ -11,7 +11,7 @@ setup(
     long_description=open("../README.md").read(),  # noqa: SIM115
     long_description_content_type="text/markdown",
     url="https://github.com/materialsproject/emmet",
-    packages=["emmet.cli"],
+    packages=["emmet.cli.legacy"],
     scripts=["emmet/cli/bash_scripts/launcher_finder"],
     install_requires=[
         "log4mongo",
@@ -31,6 +31,6 @@ setup(
     zip_safe=False,
     entry_points="""
     [console_scripts]
-    emmet=emmet.cli.entry_point:safe_entry_point
+    emmet=emmet.cli.legacy.entry_point:safe_entry_point
     """,
 )
